+++ Platform.java
@@ -5,15 +5,15 @@ import java.awt.image.BufferedImage;
 
 public class Platform extends Movable{
 
-	public Platform(BufferedImage b, int x, int y, int w, int h) {
+	public Platform(BufferedImage b, double x, double y, double w, double h) {
 		super(b, x, y, w, h, true, 0, 0);
 		isPlatform = true;
 	}
 	
 	public void draw(Graphics g) {
 		g.setColor(Color.red);
-		g.fillRect(x, y, w, h);
-		g.drawRect(x, y, w, h);
+		g.fillRect((int) x, (int) y, (int) w, (int) h);
+		g.drawRect((int) x, (int) y, (int) w, (int) h);
 	}
 
 }
